clutter-master-clock-gdk: Do not try to access invalidated list
During a stage update a stage could be removed, this was handled only once
and when it happened we did not care about updating the list of of stages
we're going through, with possible access to invalid or free'd memory.

So we while iterating and after processing the events we need to check if a
stage is still valid, if that's the case we just continue in the normal way
otherwise we ignore further stage operations and we fetch again the list
we're iterating on, in case that has changed.

See: https://autopkgtest.ubuntu.com/results/autopkgtest-lunar/lunar/amd64/c/clutter-1.0/20230308_084252_929ac@/log.gz

Gbp-Pq: Name clutter-master-clock-gdk-Do-not-try-to-access-invalidated.patch